Add cache_latency benchmark
This benchmark can be used to measure the time it takes to
access each level of cache on a system.
The idea is to create a randomized linked list where each element is a
cache line wide. A simple run through this list will generate a sequence
of access that can be predicted by the prefetcher.
Given enough runs of that list, if all elements fit in a cache level,
the time for each access will stabilize. Playing with the list size can
thus make access go to a given cache level.

This kind of experiment is studied in details in Ulrich Drepper paper :
"What every programmer should not about memory"